Release/2.0

+++ src/FubarDev.WebDavServer/Engines/Local/CollectionTarget.cs
@@ -8,29 +8,27 @@ using System.Threading.Tasks;
 
 using FubarDev.WebDavServer.FileSystem;
 
-using JetBrains.Annotations;
-
 namespace FubarDev.WebDavServer.Engines.Local
 {
     /// <summary>
-    /// The local file system collection target
+    /// The local file system collection target.
     /// </summary>
     public class CollectionTarget : EntryTarget, ICollectionTarget<CollectionTarget, DocumentTarget, MissingTarget>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="CollectionTarget"/> class.
         /// </summary>
-        /// <param name="destinationUrl">The destination URL for this collection</param>
-        /// <param name="parent">The parent collection</param>
-        /// <param name="collection">The underlying collection</param>
-        /// <param name="created">Was this collection created by the <see cref="RecursiveExecutionEngine{TCollection,TDocument,TMissing}"/></param>
-        /// <param name="targetActions">The target actions implementation to use</param>
+        /// <param name="destinationUrl">The destination URL for this collection.</param>
+        /// <param name="parent">The parent collection.</param>
+        /// <param name="collection">The underlying collection.</param>
+        /// <param name="created">Was this collection created by the <see cref="RecursiveExecutionEngine{TCollection,TDocument,TMissing}"/>.</param>
+        /// <param name="targetActions">The target actions implementation to use.</param>
         public CollectionTarget(
-            [NotNull] Uri destinationUrl,
-            [CanBeNull] CollectionTarget parent,
-            [NotNull] ICollection collection,
+            Uri destinationUrl,
+            CollectionTarget? parent,
+            ICollection collection,
             bool created,
-            [NotNull] ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
+            ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
             : base(targetActions, parent, destinationUrl, collection)
         {
             Collection = collection;
@@ -38,9 +36,8 @@ namespace FubarDev.WebDavServer.Engines.Local
         }
 
         /// <summary>
-        /// Gets the underlying collection
+        /// Gets the underlying collection.
         /// </summary>
-        [NotNull]
         public ICollection Collection { get; }
 
         /// <inheritdoc />
@@ -49,17 +46,16 @@ namespace FubarDev.WebDavServer.Engines.Local
         /// <summary>
         /// Creates a new instance of the <see cref="CollectionTarget"/> class.
         /// </summary>
-        /// <param name="destinationUrl">The destination URL for this collection</param>
-        /// <param name="collection">The underlying collection</param>
-        /// <param name="targetActions">The target actions implementation to use</param>
-        /// <returns>The created collection target object</returns>
-        [NotNull]
+        /// <param name="destinationUrl">The destination URL for this collection.</param>
+        /// <param name="collection">The underlying collection.</param>
+        /// <param name="targetActions">The target actions implementation to use.</param>
+        /// <returns>The created collection target object.</returns>
         public static CollectionTarget NewInstance(
-            [NotNull] Uri destinationUrl,
-            [NotNull] ICollection collection,
-            [NotNull] ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
+            Uri destinationUrl,
+            ICollection collection,
+            ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
         {
-            CollectionTarget parentTarget;
+            CollectionTarget? parentTarget;
             if (collection.Parent != null)
             {
                 var collUrl = destinationUrl.GetParent();
@@ -77,6 +73,11 @@ namespace FubarDev.WebDavServer.Engines.Local
         /// <inheritdoc />
         public async Task<MissingTarget> DeleteAsync(CancellationToken cancellationToken)
         {
+            if (Parent == null)
+            {
+                throw new InvalidOperationException("Cannot delete entry, because the collection is unspecified.");
+            }
+
             var name = Collection.Name;
             await Collection.DeleteAsync(cancellationToken).ConfigureAwait(false);
             return new MissingTarget(DestinationUrl, name, Parent, TargetActions);
@@ -87,11 +88,15 @@ namespace FubarDev.WebDavServer.Engines.Local
         {
             var result = await Collection.GetChildAsync(name, cancellationToken).ConfigureAwait(false);
             if (result == null)
+            {
                 return new MissingTarget(DestinationUrl.Append(name, false), name, this, TargetActions);
+            }
 
             var doc = result as IDocument;
             if (doc != null)
+            {
                 return new DocumentTarget(this, DestinationUrl.Append(doc), doc, TargetActions);
+            }
 
             var coll = (ICollection)result;
             return new CollectionTarget(DestinationUrl.Append(coll), this, coll, false, TargetActions);
